Refactor project initialization, CLI readiness, and preview process management: remove redundant asyncio usage, streamline directory operations, enhance template handling, simplify metadata storage, and optimize subprocess management. Transition billing and user credits to integer-based precision, and introduce error log retrieval for previews.

+++ apps/api/app/api/users.py
@@ -22,7 +22,7 @@ class UserProfileOut(BaseModel):
     last_login_at: Optional[datetime] = None
     last_active_at: Optional[datetime] = None
     plan: Optional[str] = None
-    credit_balance: Optional[float] = None
+    credit_balance: Optional[int] = None
 
     class Config:
         from_attributes = True
